requirements.txt
annotated-types==0.7.0
anyio==4.10.0
black==25.1.0
blis==1.3.0
catalogue==2.0.10
certifi==2025.8.3
cffi==1.17.1
cfgv==3.4.0
charset-normalizer==3.4.3
click==8.2.1
cloudpathlib==0.21.1
confection==0.1.5
cryptography==45.0.6
cymem==2.0.11
distlib==0.4.0
distro==1.9.0
exceptiongroup==1.3.0
fastapi==0.116.1
filelock==3.19.1
flake8==7.3.0
fsspec==2025.7.0
h11==0.16.0
hf-xet==1.1.9
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.34.4
identify==2.6.13
idna==3.10
isort==6.0.1
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.2
langcodes==3.5.0
language_data==1.3.0
marisa-trie==1.3.1
markdown-it-py==4.0.0
MarkupSafe==3.0.2
mccabe==0.7.0
mdurl==0.1.2
mpmath==1.3.0
murmurhash==1.0.13
mypy_extensions==1.1.0
networkx==3.4.2
nodeenv==1.9.1
numpy==2.2.6
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.3
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvtx-cu12==12.8.90
openai==1.102.0
packaging==25.0
pathspec==0.12.1
pdfminer.six==20250506
pdfplumber==0.11.7
pillow==11.0.0
platformdirs==4.4.0
pre_commit==4.3.0
preshed==3.0.10
pt_core_news_sm @ https://github.com/explosion/spacy-models/releases/download/pt_core_news_sm-3.8.0/pt_core_news_sm-3.8.0-py3-none-any.whl#sha256=c304fa04db3af73cd08a250feacf560506e15a2ec2469bd1b09f06847f6b455c
pycodestyle==2.14.0
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pyflakes==3.4.0
Pygments==2.19.2
pypdf==6.0.0
pypdfium2==4.30.0
python-dotenv==1.1.1
python-multipart==0.0.20
PyYAML==6.0.2
regex==2025.7.34
requests==2.32.5
rich==14.1.0
safetensors==0.6.2
scikit-learn==1.7.1
scipy==1.15.3
shellingham==1.5.4
smart_open==7.3.0.post1
sniffio==1.3.1
spacy==3.8.7
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.1
starlette==0.47.3
sympy==1.14.0
thinc==8.3.6
threadpoolctl==3.6.0
tokenizers==0.21.4
tomli==2.2.1
torch==2.8.0
torchaudio==2.8.0+cpu
torchvision==0.23.0+cpu
tqdm==4.67.1
transformers==4.55.4
triton==3.4.0
typer==0.16.1
typing-inspection==0.4.1
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.35.0
virtualenv==20.34.0
wasabi==1.1.3
weasel==0.4.1
wrapt==1.17.3
